requirements.txt
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
decorator==5.2.1
idna==3.10
joblib==1.3.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
msgpack==1.1.0
numba==0.60.0
numpy==1.26.4
packaging==24.2
platformdirs==4.3.7
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
requests==2.32.3
scikit-learn==1.4.0
scipy==1.13.1
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
typing_extensions==4.13.1
urllib3==2.3.0
Flask-SocketIO==5.3.6
eventlet==0.33.3  # Required for Flask-SocketIO
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio[client]==5.11.1
websocket-client==1.7.0
pyaudio==0.2.14
